+++ CHANGELOG.md
@@ -1,5 +1,15 @@
 # Changelog
 
+### v3.0.0 (2018/1/2)
+
+React's new Context API has been finalized, and it uses functional `children` rather than a prop
+named `render`. Accordingly, this library has been updated to use `children` as the default.
+
+**Breaking**
+
+- `<Composer/>` now uses `children` as the render prop, rather than `render`.
+- The default `renderPropName` is now `"children"` rather than `"render"`
+
 ### v2.1.0 (2018/1/22)
 
 **New Features**